Add files via upload
The Indian Premier League, or IPL, is a T20 cricket league, which was founded in 2008 and is held every year. It sees participation from both national and international players, and eight teams representing eight Indian cities compete with each other in a double round-robin format in the league stages, which are followed by playoffs. Over the years, IPL has become one of the most-watched and most attended live sporting events all over the world.

 

You work as a data analyst at IFP, a nationally recognised news agency, which is based out of New Delhi, and provides news reports and feeds to magazines, newspapers and TV broadcasters all over the country. The Sports Editor of the agency has approached you to build a Tableau dashboard of IPL statistics over the years since its inception in order to create an infographic for a newsletter that their team is working on. For this newsletter, in some cases, they will use the visual representations as you have created in Tableau directly for their infographic, and in a few other cases, they will use important statistics after trying out the different filters and customisations that you have provided for interactivity. Therefore, you are expected to build an interactive dashboard in Tableau for this purpose. Let’s watch the upcoming video and listen to Shreyas as he states the broad objective of this assignment.